@@ -1,12 +1,20 @@
 import React from "react";
+import { createRoot } from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
-import ReactDOM from "react-dom";
+import { ApolloProvider, ApolloClient, InMemoryCache } from "@apollo/client";
 import App from "App";
 import "style.css";
+const client = new ApolloClient({
+	uri: "http://localhost:1337/graphql",
+	cache: new InMemoryCache(),
+});
 
-ReactDOM.render(
-	<Router>
-		<App />
-	</Router>,
-	document.getElementById("root"),
+const container = document.getElementById("root");
+const root = createRoot(container);
+root.render(
+	<ApolloProvider client={client}>
+		<Router>
+			<App />
+		</Router>
+	</ApolloProvider>,
 );
