@@ -1,58 +1,100 @@
 import React, { useState } from "react";
 import { CSSTransition, SwitchTransition } from "react-transition-group";
-import three from "assets/images/three.jpg";
-
-import five from "assets/images/five.jpg";
-import six from "assets/images/six.jpg";
+import { useQuery, gql } from "@apollo/client";
+import { useEffect } from "react";
+import { MoonLoader } from "react-spinners";
 
 import editors from "components/EditorsPick/editors.module.css";
 
 import SectionHeader from "components/Reuse/SectionHeader";
 import EditorTab from "./EditorTab";
-
 import CircularBtn from "./CircularBtn";
 
+const GET_EDITORS_PICK = gql`
+	query ($var: PostFiltersInput, $size: PaginationArg) {
+		posts(filters: $var, pagination: $size) {
+			data {
+				id
+				attributes {
+					title
+					snippet
+					thumb {
+						data {
+							attributes {
+								url
+							}
+						}
+					}
+					categories {
+						data {
+							attributes {
+								IDN
+							}
+						}
+					}
+				}
+			}
+		}
+	}
+`;
+
 export default function EditorsPick() {
-	const array = [
-		{
-			id: 1,
-			author: "Donald Abua",
-			category: "Fashion",
-			date: "Jan 17, 2019",
-			src: three,
-			title: "Every Level Of Your Life Will Demand A Different You!",
-			preview: "Fashion ipsum dolor sit amet consectetur adipisicing elit. Autem corrupti ad impedit? Incidunt, corrupti sequi.",
-			isActive: true,
+	const { data, loading, error } = useQuery(GET_EDITORS_PICK, {
+		variables: {
+			var: {
+				editorsPick: {
+					eq: true,
+				},
+			},
+			size: {
+				pageSize: 3,
+			},
 		},
+	});
 
-		{
-			id: 2,
-			author: "Vegan Bake",
-			category: "Ganja",
-			date: "Dec 16, 2020",
-			src: five,
-			title: "We Are All in the Gutter, but Some of Us Are Looking at the Stars",
-			preview: "Ganja ipsum dolor sit amet consectetur adipisicing elit. Autem corrupti ad impedit? Incidunt, corrupti sequi.",
-			isActive: false,
-		},
-		{
-			id: 3,
-			author: "Ukwun Abua ",
-			category: "Drums",
-			date: "Mar 28, 2020",
-			src: six,
-			title: "That Which Does Not Kill Us Makes Us Stronger",
-			preview: "David ipsum dolor sit amet consectetur adipisicing elit. Autem corrupti ad impedit? Incidunt, corrupti sequi.",
-			isActive: false,
-		},
-	];
+	const [idd, setIdd] = useState();
+	const [picks, setPicks] = useState();
 
-	const [picks, setPicks] = useState(array);
-	const [idd, setIdd] = useState(1);
+	useEffect(() => {
+		if (!loading) {
+			setIdd(data.posts.data[0].id);
+			setPicks(tempPicks);
+		}
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [loading]);
 
-	const renderBtns = picks.map(({ id, isActive }) => (
-		<CircularBtn key={id} id={id} Fn={switchPicks} isActive={isActive} />
-	));
+	if (loading)
+		return (
+			<MoonLoader
+				cssOverride={{ margin: "0 auto" }}
+				color='var(--secondary)'
+				size={25}
+			/>
+		);
+
+	// Used on first render
+	const tempPicks = data.posts.data.map((pick, index) =>
+		!index ? { ...pick, isActive: true } : { ...pick, isActive: false },
+	);
+	const tempID = data.posts.data[0].id;
+
+	const renderBtns = !picks
+		? tempPicks?.map(({ id, isActive }) => (
+				<CircularBtn
+					key={id}
+					id={id}
+					Fn={switchPicks}
+					isActive={isActive}
+				/>
+		  ))
+		: picks?.map(({ id, isActive }) => (
+				<CircularBtn
+					key={id}
+					id={id}
+					Fn={switchPicks}
+					isActive={isActive}
+				/>
+		  ));
 
 	function switchPicks(id) {
 		setIdd(id);
@@ -65,7 +107,10 @@ export default function EditorsPick() {
 		);
 	}
 
-	const pick = picks.find((pick) => pick.id === idd);
+	// Select Post
+	const pick = !picks
+		? tempPicks.find((pick) => pick.id === tempID)
+		: picks.find((pick) => pick.id === idd);
 
 	return (
 		<section className={editors.editorsPick}>
@@ -73,23 +118,10 @@ export default function EditorsPick() {
 				text={`Editor's Pick`}
 				description='Lorem ipsum dolor sit amet adipisicing elit.'
 			/>
+
 			<SwitchTransition>
-				<CSSTransition
-					key={idd}
-					timeout={1000}
-					classNames={editors}
-				>
-					<EditorTab
-						css={editors}
-						author={pick.author}
-						category={pick.category}
-						date={pick.date}
-						title={pick.title}
-						preview={pick.preview}
-						btns={renderBtns}
-						src={six}
-						thumb={pick.src}
-					/>
+				<CSSTransition key={idd} timeout={1000} classNames={editors}>
+					<EditorTab btns={renderBtns} pick={pick} />
 				</CSSTransition>
 			</SwitchTransition>
 		</section>
